Fixed bug in form minion causing various issues. Fixed bug in engine causing problems with oepc for plugins. Fixed button on group profile page. Corrected a few typos in comment api and lib changing plug_id to plug_item_id
Added possible hacky fix for threads api and comments. I want the ability for comments on comments so I added that enum to the database and backend processing for it not completely implemented yet. This sends api handling from threads to comments so that the last plugin remains thread and entries into the database indicate "thread" for the key `plug`. This will allow the key `plug` to also use "comment" for making subcommits.

Added include for thread plugin.conf.php to show_thread.php to correct $oepc and fix sql queries. but I think this should be included automatically? I need to look into this more.